+++ cli/cli/Commands/DeploymentCommands/DeployArgs.cs
@@ -14,9 +14,11 @@ public class DeployArgs
 	public static void AddDockerComposeOutputOptions<TArgs>(AppCommand<TArgs> command)
 		where TArgs : CommandArgs, IHasDockerComposeArgs
 	{
-		command.AddOption(new Option<string>(new string[]{"--docker-compose-dir", "-dcd"}, () => "", $"Specify an output path where a docker-compose project should be created. The compose file can be used to run services locally."),
+		command.AddOption(new Option<string>(new string[]{"--docker-compose-dir", "-dcd"}, () => "", 
+				description: $"Specify an output path where a new docker-compose project will be created. " +
+				             $"The compose file can be used to run services locally. " +
+				             $"(Note, existing files in this folder will be overwritten)"),
 			(args, i) => args.DockerComposeDirectoryPath = i);
-		// TODO: consider adding remote-storage ? 
 	}
 	
 	public static void AddPlanOptions<TArgs>(AppCommand<TArgs> command)