+++ microservice/microservice/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Project references without assembly names no longer cause `NRE`
+
 ### Changed
 - Local development uses the developer's access token to authenticate the Microservice